feat: Citadel v200.31 — Final Build
- Frontend: npm run build (1986 modules, 0 errors)
- nginx: resolver 127.0.0.11 + dynamic proxy_pass for asterisk (no crash if asterisk down)
- docker-compose.prod.yml: certbot_webroot volume + /etc/letsencrypt mount restored
- scripts/certbot_issue.sh: automated SSL cert issuance script for prod server
- alembic: migrations added
- .gitignore: nginx/certs/ excluded

+++ app/models/all_models.py
@@ -338,7 +338,7 @@ class FinancialLog(Base):
     driver_id = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True, index=True)
     park_name = Column(String, default="PRO", index=True)
     entry_type = Column(String, default="auto_deduction")
-    amount = Column(Float, default=0.0)
+    amount = Column(Numeric(15, 2), default=0.0)  # Decimal precision: ни одной копейки не потеряем
     note = Column(Text, nullable=True)
     meta = Column(JSONB, default={})
     created_at = Column(DateTime, default=dt.datetime.now, index=True)
@@ -443,7 +443,7 @@ class Transaction(Base):
     contractor = Column(String)            # Контрагент
     description = Column(String)           # Описание
     plate_info = Column(String, nullable=True)  # Гос.номер (если применимо)
-    amount = Column(Float)                 # Сумма транзакции
+    amount = Column(Numeric(15, 2))        # Decimal precision: ни одной копейки не потеряем
     tx_type = Column(String, default="income")  # income, expense, transfer
     date = Column(DateTime, default=dt.datetime.now)  # Дата операции
     responsibility = Column(String, default="park")  # driver, partner, park
